+++ LineSDK/LineSDK/General/LineSDKError.swift
@@ -83,9 +83,9 @@ public enum LineSDKError: Error {
         ///
         /// Associated `APIErrorDetail`
         /// contains information about the error detail. If the response data can be converted to an `APIError` object,
-        /// it will be associated with `APIErrorDetail`. The `error` property in `APIErrorDetail` indicates the cause of an error.
-        /// Otherwise, the `detail.error` will be `nil`. In both cases, `detail.raw` and `detail.rawString` will
-        /// contain the plain response and error text respectively.
+        /// it will be associated with `APIErrorDetail`. The `error` property in `APIErrorDetail` indicates the cause
+        /// of an error. Otherwise, the `detail.error` will be `nil`. In both cases, `detail.raw` and
+        /// `detail.rawString` will contain the plain response and error text respectively.
         case invalidHTTPStatusAPIError(detail: APIErrorDetail)
     }
     
@@ -100,7 +100,7 @@ public enum LineSDKError: Error {
     /// - callbackURLSchemeNotMatching: The received `URL` object while opening the app does not match the
     ///                                 defined URL scheme. Code 3005.
     /// - invalidSourceApplication: The source application is invalid and cannot finish the authorization
-    ///                             process. Code 3006.
+    ///                             process. Not in use anymore from LINE SDK 5.2.4. Code 3006.
     /// - malformedRedirectURL: The received `URL` object while opening the app is invalid or does not
     ///                         contain necessary information. Code 3007.
     /// - invalidLineURLResultCode: The received `URL` object while opening the app has an unknown result
@@ -138,7 +138,8 @@ public enum LineSDKError: Error {
         /// The received `URL` object while opening the app does not match the defined URL scheme. Code 3005.
         case callbackURLSchemeNotMatching
         
-        /// The source application is invalid and cannot finish the authorization process. Code 3006.
+        /// The source application is invalid and cannot finish the authorization process.
+        /// Not in use anymore from LINE SDK 5.2.4. Code 3006.
         case invalidSourceApplication
         
         /// The received `URL` object while opening the app is invalid or does not
@@ -604,11 +605,10 @@ extension LineSDKError.GeneralErrorReason {
     }
 }
 
-extension Result where Error == LineSDKError {
-    init(_ throwing: () throws -> Value) {
+extension Result where Failure == LineSDKError {
+    init(catching body: () throws -> Success) {
         do {
-            let value = try throwing()
-            self = .success(value)
+            self = .success(try body())
         } catch {
             self = .failure(error.sdkError)
         }