+++ LineSDK/LineSDKTests/Networking/SessionTests.swift
@@ -232,7 +232,7 @@ class SessionTests: XCTestCase {
                 XCTFail("Should parse to final result after restarting for once.")
                 return
             }
-            XCTAssertTrue(delegate.stubs.isEmpty)
+            XCTAssertTrue(delegate.stubItems.isEmpty)
             XCTAssertEqual(value.foo, "bar")
         }
         waitForExpectations(timeout: 1, handler: nil)
@@ -292,7 +292,7 @@ class SessionTests: XCTestCase {
                 XCTFail("Should parse to final result after restarting for once.")
                 return
             }
-            XCTAssertTrue(delegate.stubs.isEmpty)
+            XCTAssertTrue(delegate.stubItems.isEmpty)
             XCTAssertEqual(value.foo, "bar")
         }
         waitForExpectations(timeout: 1, handler: nil)
@@ -319,6 +319,37 @@ class SessionTests: XCTestCase {
             XCTAssertEqual(result.value!.foo, "bar")
             expect.fulfill()
         }
+        waitForExpectations(timeout: 1, handler: nil)
+    }
+
+    func testSessionDelegateLock() {
+
+        // There should be no deadlock when sending requests from concurrent queue.
+
+        let expect = expectation(description: "\(#file)_\(#line)")
+
+        let range = 0 ..< 1000
+        let delegate = SessionDelegate()
+        let queue = DispatchQueue(label: "test", attributes: .concurrent)
+
+        let group = DispatchGroup()
+        for _ in range {
+            group.enter()
+            queue.async {
+                delegate.add(
+                    SessionTask(
+                        session: URLSession.shared,
+                        request: URLRequest(url: URL(string: "https://example.com")!)
+                    )
+                )
+                group.leave()
+            }
+        }
+
+        group.notify(queue: .main) {
+            expect.fulfill()
+        }
+
         waitForExpectations(timeout: 1, handler: nil)
     }
 }